加入弹幕发布功能

+++ cola_data/src/cola_im/command/danmaku.rs
@@ -10,20 +10,21 @@ use serde::{Deserialize, Serialize};
 ////////
 
 /// # [COMMAND] - 弹幕发布命令
+/// * `desc`: `可见范围； 0. 不可见 1. 自己 2. 朋友 3. 粉丝 4. 所有人（黑名单除外）`
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 pub struct RoomDanmakuCommand {
     pub user_id: i64,          // 用户 ID
     pub video_id: i64,         // 视频 ID
     pub danmaku_type: i16,     // 弹幕类型: 0. 滚动 1. 固定底部 2. 固定
-    pub channel_id: i16,       // 通道
+    pub channel_id: i32,       // 通道 ID
     pub content: String,       // 内容
     pub play_time: i32,        // 播放时间（ms）
     pub duration: i16,         // 停留时长
     pub color: Option<String>, // 颜色（可选）
     pub mode: i16,             // 位置
     pub add_time: i32,         // 创建时间
-    pub visibility: i16, // 可见范围； 0. 不可见 1. 自己 2. 朋友 3. 粉丝 4. 所有人（黑名单除外）
-    pub status: i16,     // 状态: 0. 审核 1. 可见
+    pub visibility: i16,       // 可见范围
+    pub status: i16,           // 状态: 0. 审核 1. 可见
 }
 
 /// # [BUILD] - 构造新弹幕
@@ -55,10 +56,10 @@ impl RoomDanmakuCommand {
             play_time: self.play_time,   // 发射时间
             duration: final_duration,    // 🚀 默认 5000 毫秒
             color: final_color,          // 🚀 默认 "#FFFFFF"
-            mode: self.mode,
-            likes: 0,      // 初始计数为 0
-            dislikes: 0,   // 初始计数为 0
-            visibility: 5, // 默认所有人可见
+            mode: self.mode,             // 模式
+            likes: 0,                    // 初始计数为 0
+            dislikes: 0,                 // 初始计数为 0
+            visibility: 5,               // 默认所有人可见
             status: self.status,
             sync_time: now_ts,
             ..Default::default()
